move all constants into their own package for use in docs
