+++ apps/discord-bot/src/components/settings/manage-account-menu.tsx
@@ -17,13 +17,19 @@ import {
   InstructionsContainer,
   EmbedMenuInstruction,
 } from "~discord-bot/components/primitives";
+import {
+  DISABLE_INDEXING_LABEL,
+  ENABLE_INDEXING_LABEL,
+  GLOBALLY_IGNORE_ACCOUNT_LABEL,
+  GRANT_CONSENT_LABEL,
+  REVOKE_CONSENT_LABEL,
+  STOP_IGNORING_ACCOUNT_LABEL,
+} from "@answeroverflow/constants";
 
 type ManageAccountMenuItemProps = {
   state: ManageAccountMenuState;
   setSettings: (settings: ManageAccountMenuState) => void;
 };
-export const REVOKE_CONSENT_LABEL = "Disable publicly showing messages";
-export const GRANT_CONSENT_LABEL = "Publicly display messages on Answer Overflow";
 
 const ToggleConsentButton = ({ state, setSettings }: ManageAccountMenuItemProps) => (
   <ToggleButton
@@ -50,8 +56,6 @@ const ToggleConsentButton = ({ state, setSettings }: ManageAccountMenuItemProps)
   />
 );
 
-export const DISABLE_INDEXING_LABEL = "Ignore account in server";
-export const ENABLE_INDEXING_LABEL = "Enable indexing of messages in server";
 const ToggleIndexingButton = ({ state, setSettings }: ManageAccountMenuItemProps) => (
   <ToggleButton
     currentlyEnabled={!state.settings.flags.messageIndexingDisabled}
@@ -76,7 +80,6 @@ const ToggleIndexingButton = ({ state, setSettings }: ManageAccountMenuItemProps
   />
 );
 
-export const GLOBALLY_IGNORE_ACCOUNT_LABEL = "Globally ignore account";
 export const GloballyIgnoreAccountButton = ({
   setState,
 }: {
@@ -106,7 +109,6 @@ export const GloballyIgnoreAccountButton = ({
   />
 );
 
-export const STOP_IGNORING_ACCOUNT_LABEL = "Stop ignoring account";
 export const StopIgnoringAccountButton = ({
   setState,
 }: {
