+++ apps/discord-bot/src/utils/constants.ts
@@ -10,6 +10,3 @@ export const LOADING_MESSAGES = [
   "Give me a moment",
   "Loading...",
 ];
-
-export const ANSWER_OVERFLOW_BLUE_HEX = "#8CD1FF";
-export const ANSWER_OVERFLOW_BLUE_AS_INT = parseInt(ANSWER_OVERFLOW_BLUE_HEX.replace("#", "0x"));
