+++ packages/constants/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@answeroverflow/constants",
+  "version": "0.1.0",
+  "main": "./index.ts",
+  "types": "./index.ts",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "clean": "rm -rf .turbo node_modules coverage",
+    "with-env": "dotenv -e ../../.env --",
+    "lint": "eslint . && tsc --noEmit",
+    "lint:fix": "eslint . --fix"
+  },
+  "dependencies": {
+    "discord-api-types": "^0.37.30",
+    "dotenv-cra": "^3.0.3",
+    "zod": "^3.20.6"
+  },
+  "devDependencies": {
+    "@answeroverflow/tsconfig": "*",
+    "dotenv-cli": "^7.0.0"
+  }
+}
